app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
requests
python-dotenv
